requirements.txt
twitchapi
torch
python-dotenv
openai
python-vlc
transformers
google-cloud-texttospeech
scipy
pyaudio
keybaoard
accelerate